+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/i18n/I18n.kt
@@ -581,6 +581,15 @@ open class I18nView(
         DE - "noch nicht verifiziert"
     }
 
+    open fun verificationAlreadyRunning() = translate {
+        EN - "There is already a verification process with this user, please finish/cancel it before starting a new verification"
+        DE - "Es läuft bereits ein Verifikationsprozess mit diesem Nutzer, bitte beenden oder brechen Sie diesen ab, bevor Sie eine neue Verifikation starten"
+    }
+
+    open fun verificationAlreadyRunningInAnotherRoom() = translate {
+        EN - "There is already a verification process with this user in another room"
+        DE - "Es läuft bereits ein Verifikationsprozess mit diesem Nutzer in einem anderen Raum"
+    }
     open fun dehydratedDevice() = translate {
         EN - "This virtual (so called dehydrated) device is used to decrypt messages when no other device is online."
         DE - "Dieses virtuelle (sogenannte dehydrierte) Gerät wird verwendet um Nachrichten zu entschlüsseln, wenn kein anderes Gerät online ist."
@@ -873,6 +882,11 @@ open class I18nView(
     open fun userProfileVerification() = translate {
         EN - "start user verification"
         DE - "Vertrauensprüfung starten"
+    }
+
+    open fun userProfileNavigateToVerification() = translate {
+        EN - "to verification process"
+        DE - "Zur Vertrauensprüfung"
     }
 
     open fun userProfileCopyUserId() = translate {